Removed old href, ihref properties.
