+++ tests/Sabre/CalDAV/Schedule/PluginPropertiesWithSharedCalendarTest.php
@@ -19,7 +19,7 @@ class PluginPropertiesWithSharedCalendarTest extends \Sabre\DAVServerTest {
             'principals/user1',
             'shared',
             [
-                '{http://calendarserver.org/ns/}shared-url' => new DAV\Property\Href('calendars/user2/default/'),
+                '{http://calendarserver.org/ns/}shared-url' => new DAV\Xml\Property\Href('calendars/user2/default/'),
                 '{http://sabredav.org/ns}read-only' => false,
                 '{http://sabredav.org/ns}owner-principal' => 'principals/user2',
             ]
